vite works and the front end looks good next step is to work on the stamp cards
