+++ src/main/java/com/fox/cradle/DatabaseInitializer.java
@@ -4,18 +4,17 @@ import com.fox.cradle.configuration.security.user.User;
 import com.fox.cradle.configuration.security.user.UserRepository;
 import com.fox.cradle.features.appuser.model.AppUser;
 import com.fox.cradle.features.appuser.service.AppUserService;
-import com.fox.cradle.features.picture.model.Picture;
 import com.fox.cradle.features.picture.service.PictureService;
-import com.fox.cradle.features.stamp.model.enums.StampCardCategory;
-import com.fox.cradle.features.stamp.model.enums.StampCardSecurity;
-import com.fox.cradle.features.stamp.model.enums.StampCardStatus;
-import com.fox.cradle.features.stamp.model.template.Template;
+import com.fox.cradle.features.stampSystem.model.enums.StampCardCategory;
+import com.fox.cradle.features.stampSystem.model.enums.StampCardSecurity;
+import com.fox.cradle.features.stampSystem.model.enums.StampCardStatus;
+import com.fox.cradle.features.stampSystem.model.template.Template;
 import com.fox.cradle.features.news.model.News;
 import com.fox.cradle.features.news.model.NewsCategory;
 import com.fox.cradle.features.news.service.NewsService;
-import com.fox.cradle.features.stamp.service.card.StampCardService;
-import com.fox.cradle.features.stamp.service.stamp.StampService;
-import com.fox.cradle.features.stamp.service.template.TemplateService;
+import com.fox.cradle.features.stampSystem.service.card.StampCardService;
+import com.fox.cradle.features.stampSystem.service.stamp.StampService;
+import com.fox.cradle.features.stampSystem.service.template.TemplateService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
