+++ src/main/java/com/fox/cradle/features/appuser/model/AppUser.java
@@ -1,6 +1,6 @@
 package com.fox.cradle.features.appuser.model;
 
-import com.fox.cradle.features.stamp.model.stampcard.StampCard;
+import com.fox.cradle.features.stampSystem.model.stampcard.StampCard;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
