--- src/main/java/com/fox/cradle/features/stamp/model/enums/StampCardCategory.java
@@ -1,9 +1,0 @@
-package com.fox.cradle.features.stamp.model.enums;
-
-public enum StampCardCategory
-{
-    FOOD,
-    DRINK,
-    ENTERTAINMENT,
-    OTHER;
-}
