--- src/main/java/com/fox/cradle/features/stamp/model/enums/StampCardStatus.java
@@ -1,10 +1,0 @@
-package com.fox.cradle.features.stamp.model.enums;
-
-public enum StampCardStatus
-{
-    ACTIVE,
-    INACTIVE,
-    DELETED,
-    PRIVATE,
-    PUBLIC,
-}
