--- src/main/java/com/fox/cradle/features/stamp/model/template/TemplateEdit.java
@@ -1,20 +1,0 @@
-package com.fox.cradle.features.stamp.model.template;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class TemplateEdit
-{
-    private String id;
-    private String name;
-    private String description;
-    private String defaultCount;
-    private String createdBy;
-    private String image;
-    private String stampCardCategory;
-    private String stampCardSecurity;
-    private String stampCardStatus;
-    private String createdDate;
-}
