--- src/main/java/com/fox/cradle/features/stamp/service/template/TemplateRepository.java
@@ -1,10 +1,0 @@
-package com.fox.cradle.features.stamp.service.template;
-
-import com.fox.cradle.features.stamp.model.template.Template;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface TemplateRepository extends JpaRepository<Template, Long>
-{
-}
