+++ src/main/java/com/fox/cradle/features/stampSystem/model/enums/StampCardCategory.java
@@ -1,0 +1,9 @@
+package com.fox.cradle.features.stampSystem.model.enums;
+
+public enum StampCardCategory
+{
+    FOOD,
+    DRINK,
+    ENTERTAINMENT,
+    OTHER;
+}
