+++ src/main/java/com/fox/cradle/features/stampSystem/model/enums/StampCardStatus.java
@@ -1,0 +1,10 @@
+package com.fox.cradle.features.stampSystem.model.enums;
+
+public enum StampCardStatus
+{
+    ACTIVE,
+    INACTIVE,
+    DELETED,
+    PRIVATE,
+    PUBLIC,
+}
