+++ src/main/java/com/fox/cradle/globalconfig/DataConfig.java
@@ -8,7 +8,7 @@ import org.springframework.data.mongodb.repository.config.EnableMongoRepositorie
 @EnableJpaRepositories(basePackages = {
         "com.fox.cradle.configuration.security",
         "com.fox.cradle.features.appuser",
-        "com.fox.cradle.features.stamp"
+        "com.fox.cradle.features.stampSystem"
 })
 @EnableMongoRepositories(basePackages = {
         "com.fox.cradle.features.picture",
